refactor: creating a folder to navigation domain
