+++ src/modules/authentication/containers/SignInScreen.js
@@ -5,7 +5,7 @@ import { connect } from 'react-redux'
 import { signInFirebase } from '../../../redux/actions'
 import { LoginForm } from '../components/LoginForm'
 import { SignInFooter } from '../components/SignInFooter'
-import Navigation from '../../../configuration/navigation'
+import Navigation from '../../../configuration/Navigation'
 
 export class SignInScreen extends Component {
   static propTypes = {
